Use eventhub mgmt client to create test eventhubs namespace and resources
