+++ sdk/eventhub/azure-eventhub/dev_requirements.txt
@@ -1,7 +1,7 @@
 -e ../../../tools/azure-sdk-tools
 ../../core/azure-core
 -e ../../identity/azure-identity
--e ../../servicebus/azure-servicebus
+-e ../azure-mgmt-eventhub
 aiohttp>=3.0; python_version >= '3.5'
 docutils>=0.14
 pygments>=2.2.0
