+++ sdk/eventhub/azure-eventhub/tests/livetest/synctests/test_auth.py
@@ -11,18 +11,15 @@ from azure.eventhub import EventData, EventHubProducerClient, EventHubConsumerCl
 
 
 @pytest.mark.liveTest
-def test_client_secret_credential(aad_credential, live_eventhub):
-    try:
-        from azure.identity import EnvironmentCredential
-    except:
-        pytest.skip("No azure identity library")
+def test_client_secret_credential(aad_credential, aad_credential_test_eh):
+    from azure.identity import EnvironmentCredential
     credential = EnvironmentCredential()
-    producer_client = EventHubProducerClient(fully_qualified_namespace=live_eventhub['hostname'],
-                                             eventhub_name=live_eventhub['event_hub'],
+    producer_client = EventHubProducerClient(fully_qualified_namespace=aad_credential_test_eh['hostname'],
+                                             eventhub_name=aad_credential_test_eh['event_hub'],
                                              credential=credential,
                                              user_agent='customized information')
-    consumer_client = EventHubConsumerClient(fully_qualified_namespace=live_eventhub['hostname'],
-                                             eventhub_name=live_eventhub['event_hub'],
+    consumer_client = EventHubConsumerClient(fully_qualified_namespace=aad_credential_test_eh['hostname'],
+                                             eventhub_name=aad_credential_test_eh['event_hub'],
                                              consumer_group='$default',
                                              credential=credential,
                                              user_agent='customized information')
@@ -47,5 +44,4 @@ def test_client_secret_credential(aad_credential, live_eventhub):
 
     worker.join()
     assert on_event.called is True
-    assert on_event.partition_id == "0"
-    assert list(on_event.event.body)[0] == 'A single message'.encode('utf-8')
+
