+++ sdk/eventhub/tests.yml
@@ -30,6 +30,7 @@ jobs:
         EVENT_HUB_NAMESPACE: $(python-eh-livetest-event-hub-namespace)
         IOTHUB_CONNECTION_STR: $(python-eh-livetest-event-hub-iothub-connection-str)
         IOTHUB_DEVICE: $(python-eh-livetest-event-hub-iothub-device)
-        AZURE_CLIENT_ID: $(python-eh-livetest-event-hub-aad-client-id)
-        AZURE_TENANT_ID: $(python-eh-livetest-event-hub-aad-tenant-id)
-        AZURE_CLIENT_SECRET: $(python-eh-livetest-event-hub-aad-secret)
+        AZURE_TENANT_ID: $(aad-azure-sdk-test-tenant-id)
+        AZURE_SUBSCRIPTION_ID: $(azure-subscription-id)
+        AZURE_CLIENT_SECRET: $(aad-azure-sdk-test-client-secret)
+        AZURE_CLIENT_ID: $(aad-azure-sdk-test-client-id)
